pyproject.toml
[project]
name = "cobrawap"
description = "Collaborative Brain Wave Analysis Pipeline (Cobrawap)"
version = "0.2.1"
readme = "README.rst"
authors = [
    { name = "Cobrawap authors and contributors", email = "contact@cobrawap.org"}
]

requires-python = ">=3.8"
dependencies = [
    "ruamel.yaml == 0.17.32",
    "jinja2 >= 2.10.3",
    "scipy >= 1.7.3",
    "pygments >= 2.4.2",
    "snakemake >= 7.10.0, < 8.0.0",
    "h5py",
    "shapely",
    "elephant >= 1.0.0",
    "neo >= 0.10.2",
    "nixio >= 1.5.3",
    "pillow >= 7.0.0",
    "pandas >= 1.2.0",
    "pulp < 2.8",
    "scikit-learn >= 1.1.0",
    "scikit-image >= 0.19.0",
    "matplotlib >= 3.5.1",
    "seaborn",
    "networkx",
]

[metadata]
description-file = "README.rst"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GPL-3.0-or-later",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[project.urls]
"Homepage" = "https://cobrawap.readthedocs.io/"
"Bug Tracker" = "https://github.com/NeuralEnsemble/cobrawap/issues"

[project.scripts]
cobrawap = "cobrawap.__main__:main"

[project.optional-dependencies]
dev = [
    "black",
    "bump2version",
    "flake8",
    "mypy",
    "pre-commit",
    "pydocstyle",
    "pylint",
    "pytest",
    "pytest-black",
    "pytest-cov",
    "pytest-flake8",
    "pytest-mypy",
    "pytest-pydocstyle",
    "pytest-pylint",
    "pytest-xdist",
]
doc = [
    "sphinx>=3.3.0",
    "numpydoc>=1.1.0",
    "sphinx-tabs>=1.3.0",
    "sphinx-argparse>=0.4.0",
    "sphinx-autodoc-typehints",
    "pygments",
    "recommonmark",
]
cwl = [

]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]
